Update skeleton.Rmd
Also, I think it could be helpful to users to add this example YAML into this doc to complement the example you have for the `tufte` pkg:

```
---
title: "An Example Using the Tufte Style"
author: "John Smith"
output:
  tint::tintHtml: 
     self_contained: TRUE
---
```

This YAML is what I was looking for when I noticed that the URL to the Rmd source was wrong. What do you think?